Refactored routing and nested layout components
