+++ app/root.tsx
@@ -2,10 +2,7 @@ import {
   StreetscapeProvider,
   Box,
   Heading,
-  Flex,
   Grid,
-  GridItem,
-  Accordion,
 } from "@nycplanning/streetscape";
 import {
   Links,
@@ -14,43 +11,19 @@ import {
   Scripts,
   ScrollRestoration,
   isRouteErrorResponse,
-  useLoaderData,
-  useLocation,
   useRouteError,
-  useSearchParams,
-  LoaderFunctionArgs,
   LinksFunction,
+  useSearchParams,
 } from "react-router";
-import { Atlas, INITIAL_VIEW_STATE } from "./components/atlas.client";
+import { FlyToInterpolator, MapViewState } from "@deck.gl/core";
 import { ClientOnly } from "remix-utils/client-only";
-
-import {
-  findBoroughs,
-  findCityCouncilDistricts,
-  findCommunityDistrictsByBoroughId,
-  findAgencyBudgets,
-  findCapitalProjectManagingAgencies,
-} from "./gen";
-import { FilterMenu } from "./components/FilterMenu";
-import { SearchByAttributeMenu } from "./components/SearchByAttributeMenu";
 import { useEffect, useState } from "react";
 import {
   initializeMatomoTagManager,
   initFullStoryAnalytics,
 } from "./utils/analytics";
-import { env } from "./utils/env";
-import { BoroughId, DistrictType } from "./utils/types";
-import { FlyToInterpolator, MapViewState } from "@deck.gl/core";
 import { HeaderBar } from "./components/HeaderBar";
-import { HowToUseThisTool } from "./components/AdminDropdownContent/HowToUseThisTool";
-import { MapLayersPanel } from "./components/AdminMapLayersPanel";
-import About from "./routes/about";
-import { useUpdateSearchParams } from "./utils/utils";
-import { CommunityBoardBudgetRequestLegend } from "./components/CommunityBoardBudgetRequestLegend";
-import {
-  CapitalProjectLayerToggle,
-  CommunityBoardBudgetRequestLayerToggle,
-} from "./components/MapLayerToggle";
+import { INITIAL_VIEW_STATE } from "./components/atlas.client";
 
 export const links: LinksFunction = () => {
   return [
@@ -60,74 +33,6 @@ export const links: LinksFunction = () => {
       type: "image/x-icon",
     },
   ];
-};
-
-const { zoningApiUrl } = env;
-
-export const loader = async ({ request }: LoaderFunctionArgs) => {
-  const url = new URL(request.url);
-  const districtType = url.searchParams.get("districtType") as DistrictType;
-  const boroughId = url.searchParams.get("boroughId") as BoroughId;
-
-  const { managingAgencies } = await findCapitalProjectManagingAgencies({
-    baseURL: `${zoningApiUrl}/api`,
-  });
-
-  const { agencyBudgets } = await findAgencyBudgets({
-    baseURL: `${zoningApiUrl}/api`,
-  });
-
-  if (districtType === "cd") {
-    const { boroughs } = await findBoroughs({
-      baseURL: `${zoningApiUrl}/api`,
-    });
-
-    if (boroughId === null) {
-      return {
-        boroughs,
-        communityDistricts: null,
-        cityCouncilDistricts: null,
-        managingAgencies,
-        agencyBudgets,
-      };
-    } else {
-      const { communityDistricts } = await findCommunityDistrictsByBoroughId(
-        boroughId,
-        {
-          baseURL: `${zoningApiUrl}/api`,
-        },
-      );
-
-      return {
-        boroughs,
-        communityDistricts,
-        cityCouncilDistricts: null,
-        managingAgencies,
-        agencyBudgets,
-      };
-    }
-  }
-
-  if (districtType === "ccd") {
-    const { cityCouncilDistricts } = await findCityCouncilDistricts({
-      baseURL: `${zoningApiUrl}/api`,
-    });
-    return {
-      boroughs: null,
-      communityDistricts: null,
-      cityCouncilDistricts,
-      managingAgencies,
-      agencyBudgets,
-    };
-  }
-
-  return {
-    boroughs: null,
-    communityDistricts: null,
-    cityCouncilDistricts: null,
-    managingAgencies,
-    agencyBudgets,
-  };
 };
 
 function Document({
@@ -160,24 +65,18 @@ function Document({
   );
 }
 
+export type RootContextType = {
+  viewState: MapViewState;
+  setViewState: (newViewState: MapViewState) => void;
+};
+
 export default function App() {
   useEffect(() => {
     initializeMatomoTagManager("SmoWWpiD");
     initFullStoryAnalytics();
   }, []);
   const [viewState, setViewState] = useState<MapViewState>(INITIAL_VIEW_STATE);
-  const [searchParams, setSearchParams] = useSearchParams();
-  const [, updateSearchParams] = useUpdateSearchParams();
-  const showCapitalProjects = searchParams.get("capitalProjects") !== "off";
-  const showCbbr = searchParams.get("cbbr") !== "off";
-
-  const {
-    boroughs,
-    communityDistricts,
-    cityCouncilDistricts,
-    managingAgencies,
-    agencyBudgets,
-  } = useLoaderData<typeof loader>();
+  const [, setSearchParams] = useSearchParams();
 
   const clearAllFilters = () => {
     setSearchParams({});
@@ -188,175 +87,39 @@ export default function App() {
     });
   };
 
-  const clearCapitalProjectFilters = () => {
-    updateSearchParams({
-      managingAgency: null,
-      agencyBudget: null,
-      commitmentsTotalMin: null,
-      commitmentsTotalMax: null,
-    });
-  };
-
-  const location = useLocation();
-
   return (
     <Document>
       <StreetscapeProvider>
         <ClientOnly>
-          {() =>
-            location.pathname === "/about" ? (
-              <About />
-            ) : (
-              <>
-                <Atlas
-                  viewState={viewState}
-                  setViewState={(MapViewState) => setViewState(MapViewState)}
-                  showCapitalProjects={showCapitalProjects}
-                  showCbbr={showCbbr}
-                />{" "}
-                <Grid
-                  templateColumns={{
-                    base: "0 [col-start] 1fr repeat(6, 1fr) 1fr [col-end] 0",
-                    md: "1.5dvw [col-start] 1fr repeat(10, 1fr) 1fr [col-end] 1.5dvw",
-                    lg: "1.18dvw [col-start] 1fr repeat(10, 1fr) 1fr [col-end] 1.18dvw",
-                    xl: "0.86dvw [col-start] 1fr repeat(10, 1fr) 1fr [col-end] 0.86dvw",
-                    "2xl":
-                      "0.8dvw [col-start] 1fr repeat(10, 1fr) 1fr [col-end] 0.82dvw",
-                  }}
-                  gap={{
-                    base: "0 3dvw",
-                    md: "0 1.6dvw",
-                    lg: "0 1.22dvw",
-                    xl: "0 0.94dw",
-                    "2xl": "0 0.78dw",
-                  }}
-                  templateRows={{
-                    base: "7dvh 2dvh [row-start] 1fr [row-end] 2dvh 7dvh",
-                    md: "7dvh 2dvh [row-start] 1fr [row-end] 2dvh",
-                  }}
-                  height="100vh"
-                  sx={{
-                    scrollbarWidth: "none",
-                  }}
-                >
-                  <HeaderBar clearSelections={clearAllFilters} />
-                  <GridItem
-                    gridColumn={{
-                      base: "col-start / span 7",
-                      md: "col-start / span 4",
-                      xl: "col-start / span 3",
-                      "2xl": "col-start / span 2",
-                    }}
-                    gridRow={{
-                      base: "row-start / row-end",
-                      md: "row-start / row-end",
-                      lg: "row-start / span 1",
-                    }}
-                    height={"100%"}
-                    overflowY={{ lg: "scroll" }}
-                    zIndex={"1"}
-                    sx={{
-                      scrollbarWidth: "none",
-                    }}
-                  >
-                    <Flex
-                      direction={"column"}
-                      width={{ base: "100%", lg: "auto" }}
-                      alignItems={"center"}
-                      flexShrink={{ lg: 0 }}
-                      maxHeight={{
-                        base: "82vh",
-                        lg: "full",
-                      }}
-                      backgroundColor={"white"}
-                      borderRadius={10}
-                      overflowY={"scroll"}
-                      padding={4}
-                      sx={{
-                        scrollbarWidth: "none",
-                      }}
-                      boxShadow={"0 2px 8px 0 rgba(0, 0, 0, 0.20)"}
-                    >
-                      <Accordion
-                        allowMultiple
-                        defaultIndex={[0]}
-                        width={"100%"}
-                      >
-                        <MapLayersPanel>
-                          <Box
-                            display={"flex"}
-                            flexDirection={"column"}
-                            gap={2}
-                          >
-                            <CapitalProjectLayerToggle />
-                            <SearchByAttributeMenu
-                              agencies={managingAgencies}
-                              projectTypes={agencyBudgets}
-                              onClear={clearCapitalProjectFilters}
-                            />
-                            <CommunityBoardBudgetRequestLayerToggle />
-                            <CommunityBoardBudgetRequestLegend />
-                          </Box>
-                        </MapLayersPanel>
-                        <FilterMenu
-                          boroughs={boroughs}
-                          communityDistricts={communityDistricts}
-                          cityCouncilDistricts={cityCouncilDistricts}
-                        />
-                        <HowToUseThisTool />
-                      </Accordion>
-                    </Flex>
-                  </GridItem>
-                  <GridItem
-                    gridColumn={{
-                      base: "1 / -1",
-                      md: "9 / span 5",
-                      xl: "10 / col-end",
-                      "2xl": "11 / col-end",
-                    }}
-                    gridRow={{
-                      base: "3 / -1",
-                      md: "row-start / row-end",
-                      lg: "row-start / span 1",
-                    }}
-                    height={"100%"}
-                    pointerEvents={"none"}
-                    zIndex={"2"}
-                    sx={{
-                      scrollbarWidth: "none",
-                    }}
-                    overflowY={"scroll"}
-                    display={"flex"}
-                    flexDirection={"column"}
-                    justifyContent={{ base: "end", md: "start" }}
-                  >
-                    <Flex
-                      width={"full"}
-                      gap={3}
-                      pointerEvents={"none"}
-                      sx={{
-                        "> *": {
-                          pointerEvents: "auto",
-                        },
-                        scrollbarWidth: "none",
-                      }}
-                      direction={"column"}
-                      flexShrink={{ lg: 0 }}
-                      maxHeight={"full"}
-                      justify={"end"}
-                      backgroundColor={"white"}
-                      borderRadius={10}
-                      overflowY={"scroll"}
-                      padding={4}
-                      boxShadow={"0 8px 4px 0 rgba(0, 0, 0, 0.08)"}
-                    >
-                      <Outlet />
-                    </Flex>
-                  </GridItem>
-                </Grid>
-              </>
-            )
-          }
+          {() => (
+            <Grid
+              templateColumns={{
+                base: "0 [col-start] 1fr repeat(6, 1fr) 1fr [col-end] 0",
+                md: "1.5dvw [col-start] 1fr repeat(10, 1fr) 1fr [col-end] 1.5dvw",
+                lg: "1.18dvw [col-start] 1fr repeat(10, 1fr) 1fr [col-end] 1.18dvw",
+                xl: "0.86dvw [col-start] 1fr repeat(10, 1fr) 1fr [col-end] 0.86dvw",
+                "2xl":
+                  "0.8dvw [col-start] 1fr repeat(10, 1fr) 1fr [col-end] 0.82dvw",
+              }}
+              gap={{
+                base: "0 3dvw",
+                md: "0 1.6dvw",
+                lg: "0 1.22dvw",
+                xl: "0 0.94dw",
+                "2xl": "0 0.78dw",
+              }}
+              templateRows={{
+                base: "7dvh 2dvh [row-start] 1fr [row-end] 2dvh 7dvh",
+                md: "7dvh 2dvh [row-start] 1fr [row-end] 2dvh",
+              }}
+              height="100vh"
+            >
+              <HeaderBar clearSelections={clearAllFilters} />
+              <Outlet
+                context={{ viewState, setViewState } satisfies RootContextType}
+              />
+            </Grid>
+          )}
         </ClientOnly>
       </StreetscapeProvider>
     </Document>
